Move struct marshaling code to its own package
